changed 3.1 so it now gives the total sleepduration per id
changed the regression for 3.2 and 3.3 so they dont give rsquared of 0 now. added plots to confirm the trend/no trend

adjusted 3.4 to not be depended on 3.2 and 3.3

made 3.6, added a graph to visualize the weather in Chicago and added the scatterplots with ols for intensity vs weather.

also put all of the output that was previously printed in the terminal in seperate .txt files